+++ Cedid.AkilliLojistik/aspnet-core/src/Cedid.AkilliLojistik.Web/Pages/Services/MaterialModal.cshtml.cs
@@ -91,19 +91,19 @@ namespace Cedid.AkilliLojistik.Web.Pages.Services
             public int StockCodeId { get; set; }
             [SelectItems(nameof(Units))]
             public int Unit { get; set; }
-            public float Quantity { get; set; }
-            public float Price { get; set; }
-            public float Amount { get; set; }
+            public int Quantity { get; set; }
+            public decimal Price { get; set; }
+            public decimal Amount { get; set; }
             public bool IsKDVIncluded { get; set; }
-            public float KDV { get; set; }
-            public float KDVAmount { get; set; }
+            public int KDV { get; set; }
+            public decimal KDVAmount { get; set; }
             [Range(0, 100)]
-            public float Discount { get; set; }
-            public float DiscountAmount { get; set; }
+            public int Discount { get; set; }
+            public decimal DiscountAmount { get; set; }
             [Range(0, 100)]
-            public float DiscountTwo { get; set; }
-            public float DiscountTwoAmount { get; set; }
-            public float NetAmount { get; set; }
+            public int DiscountTwo { get; set; }
+            public decimal DiscountTwoAmount { get; set; }
+            public decimal NetAmount { get; set; }
             [SelectItems(nameof(WareHouseCodes))]
             public int? WareHouseCodeId { get; set; }
             [TextArea(Rows = 4)]
